Optimize Vertex type
- The application now only supports having 'Int.MAX_VALUE' elememts
  catched, since it is using collections from the Scala stdlib. This
  will prevent downcasting the vertex IDs from 'Long' to 'Int' every
  time a Vertex is generated.

+++ src/main/scala/gnormalizer/Vertex.scala
@@ -2,5 +2,5 @@ package gnormalizer
 
 object Vertex {
 
-  type Vertex = Long
+  type Vertex = Int
 }
